filter name option move to a new file
